requirements.txt
streamlit>=1.20.0
gdown==4.6.6
tensorflow==2.20.0
keras==3.12.0
numpy
h5py
opencv-python-headless
pillow
pandas
scikit-learn
matplotlib
protobuf>=4.21.0
requests
